Add interaction step ids and values to redis cache
Note: To use a hash in redis, the db result (array of dictionaries) have 
to be converted to an array (of dictionary values). To pass the values 
to the front-end, the redis result (dictionary) has to be converted back 
into an array of dictionaries.

To think about: is there another redis data structure that requires less 
data prepping?