Await UUID checks, return immediately in case of missing data, refactor
The code was refactored to remove context from the auth-helpers. In case
of authentication failure they now throw an error with the appropriate
error message, and in case of success they return a user object.

Invoking the done callback with the correct arguments is now the sole
responsibility of the `new LocalStrategy` callback argument.