More Consistent Suspended Calculation. Fix A Test"
We opted to rely on getHighestRole for the list of roles. It's performance
is pretty poor, but the number of roles is always low. It's consistent and
keeps all the logic in that permissions.js file. If we change the way roles
work, we only have to update that one function.