Fix polling on the todos screen
This fixes polling on the texter todos screen, making it so that
the information is reloaded every 5 seconds as long as the screen
is up.  This seems to have been broken by an older commit.

The fix here is to simply ignore the user's `cacheable` property,
since (1) it's hard-coded to false right now, and (2) it doesn't
matter.  The data here, by definition, needs to be queried every
time so that we can see a fairly "live" view of the text messages
that need to be dealt with.