Remove the scroll bars in every cell in the incoming-messages table
The incoming-messages table was set so that every cell had `overflow: scroll`
set, meaning that every cell would have both horizontal and vertical scroll
bars, even when they weren't necessary.

The styles also had `text-overflow: ellipsis`, so there's really no reason
to force scrolling in the cell when the text will be cut off appropriately.

For all of these `overflow: scroll` styles, I've replaced them with
`overflow: hidden`, as it is in a couple other places in that screen.