Ensure that interactions are copied deterministically
It looks like the "id" is the most important thing when trying to
order interaction steps, and if that's the case, then we have to
copy them in the correct order.  Leaving out `orderBy` lets the
database return them in any order, and we don't know what that will
be.