Copy all relevant campaign properties when copying a campaign
I typically segment my data a bit and then copy the same campaign
over and over, changing the list and the intro text a bit each time,
but everything else generally stays the same.

I noticed some issues when copying a campaign, however:

1. There is no notification that the "Copy Campaign" button did
   anything; I mashed it many times before I realized what was happening.
2. Nothing is copied from "Basics" except the title.
3. Nothing is copied from "Dynamic Assignment".
4. Nothing is copied from "Texter Experience".
5. Nothing is copied from Texting Hours.
6. The questions and responses are not copied in the right order,
   but the order seems close to the correct order at times.

This patch addresses all of those issues.  To the best of my knowledge,
all of a campaign's configuration is copied over to the new campaign.

In addition, we show a "confirm" box upon successful copy completion
offering to take the user to the edit screen for the new campaign.
If she accepts, then we'll re-route to that screen.  If she does not,
then she'll stay where she is, but at least she'll know that the
copy operation succeeded.