Include "Sent by User <userId>" in review mode
Add link to account info page for sender if currentUser is an admin in "review mode" (as indicated by URL param).

For now, just using user ID and a link to the account page to get sender's name as needed. We could add a query for user name but this will accomplish the goal of showing if a conversation was re-assigned.