New Nation Builder event RSVP action handler
Adds new `nationbuilder-rsvp` handler that polls NB for recent events
and makes them available for RSVP'ing as an action during an
interaction. An event is chosen as an action for an answer, allowing
multiple events to be RSVP'ed to as different answers.

Requires an existing NB instance and API key, and `nationbuilder_id`
to be imported as a contact custom field for each contact.

Code originally based on that of @bxjx