Fix CampaignContactsChoiceForm html rendering
Since some contact loader components are now wrapped in WithSetTheme, an Enzyme ShallowWrapper can no longer provide the required context for CampaignContactsChoiceForm. This causes an error with the CampaignContactsChoiceForm test when html() is called on the entire wrapper. Using Enzyme's mount() method for full DOM rendering can provide the required context, but it necessitates the entire test suite to be switched from the default Node environment to Jest environment. Switching to a Jest environment will break the tests which make real server calls (the tests that call processContactLoad). The solution implemented in this PR is to update the problematic test to remove the WithSetTheme child node before running html() on the wrapper. The purpose of this test is to ensure that the Save button is present, so it is not affected by removal of the WithSetTheme component.