Fix texting view broken when dynamic assignment enabled
When a texter views the texting interface and dynamic assignment is
enabled on a campaign they are assigned to, a null error is thrown
and the client view remains blank. This fixes the issue by adding a
null check.

Regression introduced in cb4c2a5792ce91dab7d582c41b637c49a4655676