Fix texting view broken when dynamic assignment enabled
When a texter views the texting interface and dynamic assignment is
enabled on a campaign they are assigned to, a null error is thrown
and the client view remains blank. This fixes the issue by adding a
null check.

Regression introduced in 9cb55da05f9cb029a6141b2e51e6c0b88354e6ce